Update README.md
Because of the rate of flux in the Console UI, we should try to refrain from providing screenshots and/or needlessly detailed instructions for every phase of the process. These things quickly get out of date, and that is potentially more confusing to developers than having slightly more general instructions like "Copy the generated API key and and paste it..." etc.